// packages/nextjs/lib/generated/ursset.ts
export const deployment = {
  chainId: 46630,
  deployBlock: 131494754,
  addresses: {
    KYCRegistry: "0x34f4cbfaa1b0c2ea6285414d6f65bc0b292361f1",
    MockIDR: "0x0f1a07a06565a008c13cf44b4268b0fa3f6ce67d",
    PropertyFactory: "0x9f50dbb80064f8c61c1cc47c576acc1a0bb54325",
  },
} as const;